Ignore some golint warnings of struct field convention
These fields' name have been exported, so if it renames those that will be a
breaking change. So this commit makes that to ignore them.

fields should be ignored:

- `ResourceUrl`
- `AccessKeyId`
- `OperatorId`

Unfortunately, golint doesn't (and probably won't) have a feature to ignore the linting line
by line, so it ignores by golint command execution.